Index the PDCP ul/dl metrics by QCI as well as bearer for O-RAN KPM.
(This is memory-wasteful, but cheapest to do this way.  Also, unfortunate
that the PDCP did not have a notion of bearer QCI, so we have to broaden
the RRC->PDCP API a little bit.  The alternative would be to call down into
the RRC to annotate the metrics, but the risk there is that the bearer might
have already gone away by that point, and then there is locking to consider,
and we want to be lock-free.  No great solutions... better to report the
ul/dl metrics from the RRC, IMO.)