RIC agent now sends real KPM data; subscriptions/data binned to periods.
KPM subscriptions are for metrics over discrete periods, and multiple
consumers can subscribe to each period, so I added a timer prio queue
that dispatches metric/indication callbacks if any consumers are
subscribed.  The timer queue has a dedicated dispatch thread, but
callbacks are run on the main agent thread.  Locking is a little
coarse in send_indications, but there's no need to improve it yet.