Update worksharing model and add option to disable it
Previously, the sliced scheduler metric would prioritize the slice that
owned the current subframe, but allow UEs that were not yet bound to a
slice (or bound to another slice) to consume leftover PRBs in the same
subframe. This was necessary to allow new UEs to associate to the eNB,
since some RRC messages (e.g. msg4) carried by the PDSCH in the
association process might not otherwise be scheduled in time to complete
the process.

The main issue with this approach was that new UEs (those not yet bound
to a slice) could fail to associate when a slice was consuming most of
the DL PRBs for every subframe, e.g., when using iperf at high
bandwidths on a sliced link. This could occur even when the UE trying to
associate belonged to a slice, since UEs aren't actually bound to slices
until most of the association process is completed. It also made it
impossible to see the effects of slicing with only a single UE, e.g.,
when using the simulated RF mode in srsLTE.

This commit updates the worksharing model in the following ways:

- It includes a periodic subframe that prioritizes the scheduling of new
  UEs over those already bound to slices in order to allow new UEs to
  associate. The periodicity can be configured, or this can be disabled,
  via a CLI argument.

- It allows worksharing outside of the subframe mentioned above to be
  disabled via another CLI argument, which in turn makes the effects of
  slicing visible in cases where only a single UE has associated to the
  eNB.