Fix reconnect bug with srslte multiqueue handler.
Turns out that when you clear a queue id, not only does it flush all
items, but it also sets the queue to not running, so in our case,
the queue no longer accepts tasks -- e.g. the next task.  This is
not the ideal implementation for us and we can't exactly override
it, so just follow the erase with an immediate call to obtain a new
agent_queue_id.  Anyway, this was blocking second and further
reconnect attempts.

Also, only set RIC_ESTABLISHED upon successful E2setupResponse.